% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_methods.R
\name{ame_plot_heatmap}
\alias{ame_plot_heatmap}
\title{Plot AME heatmap clustered by similarity in detected motifs}
\usage{
ame_plot_heatmap(
  ame,
  id = motif_id,
  group = NULL,
  value = -log10(adj.pvalue),
  group_name = NULL
)
}
\arguments{
\item{ame}{ame results data.frame}

\item{id}{column of motif ids to use (default: motif_id).}

\item{group}{grouping column if comparing across multiple ame runs (optional, default: NULL).}

\item{value}{value to display as heatmap intensity. Default: -log10(adj.pvalue). Takes function or column name as input.}

\item{group_name}{when group = NULL, name to use for input regions. Ignored if group is set.}
}
\value{

}
\description{
Plot AME heatmap clustered by similarity in detected motifs
}
\examples{
\dontrun{
results <- runAme()
ame_plot_heatmap(results)
}
}
