% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{as_universalmotif}
\alias{as_universalmotif}
\title{Convert special data.frames back into universalmotifs}
\usage{
as_universalmotif(data)
}
\arguments{
\item{data}{a universalmotif_dataframe (output from
`as_universalmotif_dataframe()`, or `runDreme()`)}
}
\value{
universalmotif list from motifs, updated to reflect the data.frame
  column values.
}
\description{
Convert special data.frames back into universalmotifs
}
\examples{
df <- universalmotif::create_motif() \%>\%
  as_universalmotif_dataframe() \%>\%
  dplyr::mutate(altname = "new_alt_name")

motifs <- as_universalmotif(df)
}
