% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreme.R
\name{dreme_nmotifs_found}
\alias{dreme_nmotifs_found}
\title{Return number of discovered DREME motifs}
\usage{
dreme_nmotifs_found(processx_out)
}
\arguments{
\item{processx_out}{output of processx run}
}
\value{
`integer(1)` of number of motifs passing threshold
}
\description{
Return number of discovered DREME motifs
}
