% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ame}
\alias{example_ame}
\title{Example runAme() output}
\format{A list object of AME results data.frames
\describe{
 \item{Increasing}{`runAme()` Results object for Increasing sites vs Static sites}
 \item{Decreasing}{`runAme()` Results object for Decreasing sites vs Static sites}
}}
\usage{
example_ame
}
\description{
Result when running AME using 100bp window around `example_chip_summits` for
"Increasing" and "Decreasing" sites, using "Static" as background.
}
\examples{
# Data can be combined into 1 large data.frame using:
# where the "behavior" column will hold the "Increasing"/"Decreasing" information
dplyr::bind_rows(example_ame, .id = "behavior")
}
\keyword{datasets}
