% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ame_large}
\alias{example_ame_large}
\title{runAme() output for example_chip_summits split by binding description}
\format{
a list of runAme() results data.frames
}
\usage{
example_ame_large
}
\description{
AME was run for "ectopic", "entopic", and "orphan" sites using shuffled background.
}
\details{
see `vignette("integrative_analysis", package = "memes")` for details.
}
\examples{
# Data can be combined into 1 large data.frame using:
dplyr::bind_rows(example_ame_large, .id = "binding_type")
}
\keyword{datasets}
