% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dreme}
\alias{example_dreme}
\title{Example runDreme() output}
\format{a data.frame}
\usage{
example_dreme
}
\description{
Result when running dreme using 100bp window around example_chip_summits
using "Decreasing" sites as foreground, and "Static" sites as background.
}
\keyword{datasets}
