% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dreme_by_binding}
\alias{example_dreme_by_binding}
\title{runDreme() output for example_chip_summits split by binding description}
\format{a runDreme results data.frame}
\usage{
example_dreme_by_binding
}
\description{
See vignette("integrative_analysis", package = "memes") for details
}
\keyword{datasets}
