% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dreme_by_sens_vs_static}
\alias{example_dreme_by_sens_vs_static}
\title{runDreme() output for example_chip_summits split by response behavior}
\format{a runDreme results data.frame}
\usage{
example_dreme_by_sens_vs_static
}
\description{
See `vignette("integrative_analysis", package = "dremeR")` for details
}
\keyword{datasets}
