% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fimo}
\alias{example_fimo}
\title{Example runFimo() output}
\format{A GRanges object of E93 motif positions within 100bp windows of `example_chip_summits`}
\usage{
example_fimo
}
\description{
Run using 100bp windows around `example_chip_summits`, using E93 motif as database.
}
\keyword{datasets}
