% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_peaks}
\alias{example_peaks}
\title{Example ChIP-seq peaks}
\format{An object of class \code{GRanges} of length 10.}
\source{
https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE141738&format=file&file=GSE141738%5Funion%5Fchip%5Fpeaks%5Fannotated%2Ecsv%2Egz
}
\usage{
example_peaks
}
\description{
10 ChIP-seq peaks from GSE141738
}
\details{
A small number of transcription factor ChIP-seq peaks as a GRanges object,
taken from [GSE141738](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE141738)
}
\keyword{datasets}
