% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tomtom}
\alias{example_tomtom}
\title{Example runTomTom() output}
\format{
a data.frame
}
\usage{
example_tomtom
}
\description{
Result when running `runTomTom(example_dreme$motif)` using FlyFactorSurvey as database
}
\keyword{datasets}
