% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xml.R
\name{importDremeXML}
\alias{importDremeXML}
\title{Import Dreme output from previous run}
\usage{
importDremeXML(dreme_xml_path)
}
\arguments{
\item{dreme_xml_path}{path to dreme.xml file}
}
\value{
data.frame with statistics for each discovered motif. The `motifs`
  column contains a universalmotif object representation in PCM format of
  each DREME motif. If no motifs are discovered, returns NULL.
}
\description{
Import Dreme output from previous run
}
\examples{
dreme_xml <- system.file("extdata/dreme.xml", package = "memes")
importDremeXML(dreme_xml)
}
\seealso{
[runDreme()]
}
