% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo.R
\name{importFimo}
\alias{importFimo}
\title{Import FIMO results}
\usage{
importFimo(fimo_tsv)
}
\arguments{
\item{fimo_tsv}{path to fimo.tsv output file}
}
\value{
GenomicRanges object for each match position. Note unless coordinates
  are genomic positions, each `seqnames` entry will be the fasta header, and
  start/end will be the position within that sequence of the match.
}
\description{
Import FIMO results
}
\examples{

\dontrun{
# TODO: example fimo.tsv shipped in extdata
importFimo("fimo.tsv")
}
}
