% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xml.R
\name{importTomTomXML}
\alias{importTomTomXML}
\title{Import tomtom data from previous run}
\usage{
importTomTomXML(tomtom_xml_path)
}
\arguments{
\item{tomtom_xml_path}{path to tomtom.xml}
}
\value{
will return data.frame with input motifs & results for best match.
  `tomtom` list column contains full tomtom data for each input motif.
  NOTE: if tomtom detects no matches for any input motif, currently will
  throw warning & return NA values for `tomtom`, `best_match_id`, and
  `best_match_motifs`.
}
\description{
Import tomtom data from previous run
}
\examples{
\dontrun{
importTomTomXML("path/to/tomtom.xml")
}
}
