% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{is_universalmotif_dataframe}
\alias{is_universalmotif_dataframe}
\title{Check whether input is a universalmotif data.frame}
\usage{
is_universalmotif_dataframe(res)
}
\arguments{
\item{res}{data.frame}
}
\value{
TRUE or FALSE
}
\description{
Check whether input is a universalmotif data.frame
}
