% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meme_is_installed}
\alias{meme_is_installed}
\title{Returns logical vector indicating valid MEME-Suite install status}
\usage{
meme_is_installed(path = NULL)
}
\arguments{
\item{path}{optional path to "meme/bin/"}
}
\value{
`logical(1)` indicating whether meme is installed with all supported utilities
}
\description{
Checks for a valid meme install using same heirarchy as `check_meme_install()`.
Returns `TRUE` if all supported utilities are found in the meme install
location, `FALSE` if not.
}
\details{
The search heirarchy is as follows:
 1. the `meme_path` function argument if set
 2. the `meme_bin` option
 3. the `MEME_BIN` environment variable
 4. the default location at `~/meme/bin`
}
\examples{
meme_is_installed()
}
\seealso{
search_meme_path check_meme_install
}
