% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_utils.R
\name{mutate_motif}
\alias{mutate_motif}
\title{Update motif columns by data values}
\usage{
mutate_motif(.data, ..., .motif = "motif", .override = NULL)
}
\arguments{
\item{.data}{data.frame of results}

\item{...}{quoted name-value pairs of columns to edit. format: "motif_column_name" = "data_column_name"}

\item{.motif}{column of universalmotif objects to edit (default: "motif")}

\item{.override}{named list or character vector where names are
motif_column_name, values are data_column_name, used instead of ... (useful for programatically passing values)}
}
\value{
`.data` where `.motif` entries have been updated to values passed to `...` or `.override`
}
\description{
**NOTE** ignores NA values
}
\details{
**NOTE** This feature is experimental and subject to change based on
user-feedback. Please provide feedback at
\link{https://github.com/snystrom/dremeR/issues/31}
}
\examples{
motif <- universalmotif::create_motif()
df <- universalmotif_to_meme_df(motif)
df <- dplyr::mutate(df, id = "newName")
df <- dplyr::mutate(df, name = "newName")
df <- mutate_motif(df, "name" = "name")
# renamed motif
df$motif
}
