% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outdir_name}
\alias{outdir_name}
\title{Helper for writing unique directory names}
\usage{
outdir_name(input, control)
}
\arguments{
\item{input}{input file path}

\item{control}{control file path}
}
\value{
directory name in the style of: "input_vs_output".
}
\description{
Helper for writing unique directory names
}
\examples{
outdir_name("condition1.fa", "backgroundSequence.fa")
}
