% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{parseTomTom}
\alias{parseTomTom}
\title{Return tomtom stat information and PWMs for matched motifs}
\usage{
parseTomTom(tomtom_xml_path)
}
\arguments{
\item{tomtom_xml_path}{path to tomtom.xml output}
}
\value{
data.frame:
    - id: name of query PWM
    - alt: alternate name of query PWM
    - match_id: name of matched PWM
    - match_alt: alt name of matched PWM
    - match_pvalue: p-value of match
    - match_evalue: E-value of match
    - match_qvalue: q-value of match
    - match_offset: number of letters the query was offset from the target match
    - match_strand: whether the motif was found on input strand (+) or as reverse-complement (-)
    - db_name: database source of matched motif
    - match_motif: universalmotif object containing the PWM that was matched
    Returns NULL if not matches detected
}
\description{
Return tomtom stat information and PWMs for matched motifs
}
\examples{
\dontrun{
parseTomTom("tomtom.xml")
}
}
