% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_utils.R
\name{update_best_match}
\alias{update_best_match}
\title{Update best match info by ranking of tomtom data}
\usage{
update_best_match(res)
}
\arguments{
\item{res}{results from runTomTom}
}
\value{
`res` with updated best_* columns
}
\description{
This function updates the best_match columns based on the rankings on the tomtom list data.
}
