% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_utils.R
\name{update_motifs}
\alias{update_motifs}
\title{Update the `motif` column to data.frame values}
\usage{
update_motifs(.data)
}
\arguments{
\item{.data}{data.frame with `motif` column}
}
\value{
.data where `motif` column has been updated to reflect the values
  from columns sharing names with unprotected universalmotif slots.
}
\description{
Update the `motif` column to data.frame values
}
\details{
**NOTE** that `consensus`, `alphabet`, `multifreq`, and `icscore` are protected columns and
cannot be updated.
Table of values updated
| `motif`  | `data.frame` |
|:--------:|:------------:|
| name     | id           |
| altname  | alt          |
| family   | family       |
| organism | organism     |
| strand   | strand       |
| nsites   | nsites       |
| bkgsites | bkgsites     |
| pval     | pval         |
| qval     | qval         |
| eval     | eval         |
}
\examples{
motif <- universalmotif::create_motif()
df <- universalmotif_to_meme_df(motif)
df <- dplyr::mutate(df, id = "newName")
df <- update_motifs(df)
# renamed motif
df$motif
}
