% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_fasta}
\alias{write_fasta}
\title{Write fasta file from stringset}
\usage{
write_fasta(seq, path = tempfile(fileext = ".fa"))
}
\arguments{
\item{seq}{`Biostrings::XStringSet`}

\item{path}{path of fasta file to write (default: temporary file)}
}
\value{
path to created fasta file
}
\description{
Write fasta file from stringset
}
\examples{
seq <- universalmotif::create_sequences()
\dontrun{
write_fasta(seq)
}
}
