% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_methods.R
\name{ame_plot_heatmap}
\alias{ame_plot_heatmap}
\title{Plot AME heatmap clustered by similarity in detected motifs}
\usage{
ame_plot_heatmap(
  ame,
  id = motif_id,
  group = NULL,
  value = -log10(adj.pvalue),
  group_name = NULL
)
}
\arguments{
\item{ame}{ame results data.frame}

\item{id}{column of motif ids to use (default: motif_id).}

\item{group}{grouping column if comparing across multiple ame runs (optional, default: NULL).}

\item{value}{value to display as heatmap intensity. Default:
-log10(adj.pvalue). Takes function or column name as input. If set to
"normalize", will use normalized rank within `group` as the heatmap values.
**If in doubt**, prefer the -log10(adj.pvalue) plot. (See "Normalized rank
visualization" section below for more details on how to interpret these
data)}

\item{group_name}{when group = NULL, name to use for input regions. Ignored if group is set.}
}
\value{

}
\description{
Plot AME heatmap clustered by similarity in detected motifs
}
\details{
Normalized rank visualization
**NOTE:** The normalized rank visualization eliminates all real values
related to statistical significance! Instead, this visualization represents
the relative ranks of hits within an AME run, which already pass a
significance threshold set during `runAME()`. This means that even if several
motifs have similar or even identical pvalues, their heatmap representation
will be a different color value based on their ranked order in the results
list. This also means that using the normalized rank visualization will
be misleading if there are only a few AME hits; it is only worth using if the
number of hits is very large (>100). Both visualizations can be useful and reveal
different properties of the data to the user.  Use `ame_compare_heatmap_methods()` to
help assess differences in the two visualizations. **If in doubt**, prefer
the `-log10(adj.pvalue)` representation.

Common mistake: if `value` is set to a string that is not "normalize", it
will return: "Error: Discrete value supplied to continuous scale". To use a
column by name, do not quote the column name.
}
\examples{
\dontrun{
results <- runAme()
ame_plot_heatmap(results)
}
}
