% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame_methods.R
\name{ame_plot_heatmap}
\alias{ame_plot_heatmap}
\title{Plot AME heatmap clustered by similarity in detected motifs}
\usage{
ame_plot_heatmap(
  ame,
  id = motif_id,
  group = NULL,
  value = -log10(adj.pvalue),
  group_name = NULL
)
}
\arguments{
\item{ame}{ame results data.frame}

\item{id}{column of motif ids to use (default: motif_id).}

\item{group}{grouping column if comparing across multiple ame runs (optional, default: NULL).}

\item{value}{value to display as heatmap intensity. Default:
-log10(adj.pvalue). Takes function or column name as input. If set to
"normalize", will use normalized rank within `group` as the heatmap values.
This can be a more appropriate visualization for between-group comparisons.}

\item{group_name}{when group = NULL, name to use for input regions. Ignored if group is set.}
}
\value{

}
\description{
Plot AME heatmap clustered by similarity in detected motifs
}
\details{
Common mistake: if `value` is set to a string that is not "normalize", it
will return: "Error: Discrete value supplied to continuous scale". To use a
column by name, do not quote the column name.
}
\examples{
\dontrun{
results <- runAme()
ame_plot_heatmap(results)
}
}
