% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{as_universalmotif}
\alias{as_universalmotif}
\title{Convert universalmotif data.frames back into universalmotifs}
\usage{
as_universalmotif(data)
}
\arguments{
\item{data}{a universalmotif_dataframe (output from
`as_universalmotif_dataframe()`, or `runDreme()`)}
}
\value{
universalmotif list from motifs, updated to reflect the data.frame
  column values.
}
\description{
Convert universalmotif data.frames back into universalmotifs
}
\examples{
motif <- universalmotif::create_motif()
df <- as_universalmotif_dataframe(motif)
df <- dplyr::mutate(motif_df, altname = "new_alt_name")

motifs <- as_universalmotif(df)
}
