% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{as_universalmotif_dataframe}
\alias{as_universalmotif_dataframe}
\title{Convert universalmotif to data.frame with motif column}
\usage{
as_universalmotif_dataframe(motif, na.rm = FALSE)
}
\arguments{
\item{motif}{universalmotif object or list of universalmotifs}
}
\value{
data.frame with all motif metadata + `motif` column containing the universalmotif object
}
\description{
Convert universalmotif to data.frame with motif column
}
\examples{
motif <- universalmotif::create_motif()
motif_df <- as_universalmotif_dataframe(motif)
}
