% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_meme_install}
\alias{check_meme_install}
\title{Check user's MEME install}
\usage{
check_meme_install(meme_path = NULL)
}
\arguments{
\item{meme_path}{path to "meme/bin" (if unset will search `MEME_BIN` environment variable or `meme_bin` option)}
}
\value{
message indicating which MEME utilities are installed and their location on disk
}
\description{
Checks if "meme/bin" directory exists, returns green check if yes, or red X if no. Returns the path it searched for.
}
\details{
If 'meme/bin' is detected, next checks if each utility is installed, green check if yes, red X if no.
}
\examples{
check_meme_install()
}
