% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{importAme}
\alias{importAme}
\title{Parse ame output}
\usage{
importAme(path, method = "fisher", sequences = FALSE)
}
\arguments{
\item{path}{path to ame results file ("ame.tsv")}

\item{method}{ame run method used (one of: c("fisher", "ranksum", "dmhg3",
"dmhg4", "pearson", "spearman")). Default: "fisher".}

\item{sequences}{FALSE or path to sequences file (only valid for method = "fisher")}
}
\value{
data.frame with method-specific results. See [AME
  results](http://meme-suite.org/doc/ame-output-format.html) webpage for more
  information. If sequences is set to a path to the sequences.tsv and method
  = "fisher", the list-column `sequences` will be appended to resulting
  data.frame.
}
\description{
Parse ame output
}
