% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimo.R
\name{importFimo}
\alias{importFimo}
\title{Import FIMO results}
\usage{
importFimo(fimo_tsv)
}
\arguments{
\item{fimo_tsv}{path to fimo.tsv output file}
}
\value{

}
\description{
Import FIMO results
}
\examples{

\dontrun{
importFimo("fimo.tsv")
}
}
