% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meme_is_installed}
\alias{meme_is_installed}
\title{Returns logical vector indicating valid MEME-Suite install status}
\usage{
meme_is_installed(path = NULL)
}
\arguments{
\item{path}{optional path to "meme/bin/"}
}
\value{
`logical(1)` indicating whether meme is installed with all supported utilities
}
\description{
Checks for a valid meme install using same heirarchy as `search_meme_path()`.
Returns `TRUE` if all supported utilities are found in the meme install
location, `FALSE` if not.
}
\examples{
# Will return TRUE if "meme/bin/" is detected
meme_is_installed()
\dontrun{
# will throw error if path to meme is invalid
meme_is_installed("bad/path")
}
}
\seealso{
search_meme_path check_meme_install
}
