% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memes-package.R
\docType{package}
\name{memes-package}
\alias{memes}
\alias{memes-package}
\title{memes: An R Interface to the MEME Suite}
\description{
Provides functions for performing motif analysis using the MEME suite as a backend.
    Functions & outputs are designed for easy integration into tidy workflows for rapid iteration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://snystrom.github.io/memes/}
  \item \url{https://github.com/snystrom/memes}
  \item Report bugs at \url{https://github.com/snystrom/memes}
}

}
\author{
\strong{Maintainer}: Spencer Nystrom \email{snystrom@email.unc.edu} (\href{https://orcid.org/0000-0003-1000-1579}{ORCID}) [copyright holder]

}
\keyword{internal}
