% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_utils.R
\name{nest_tomtom}
\alias{nest_tomtom}
\title{Nest TomTom results columns into a data.frame column named "tomtom"}
\usage{
nest_tomtom(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Nest TomTom results columns into a data.frame column named "tomtom"
}
\examples{
\dontrun{
#TODO: add better example
res <- runTomTom(motifs)
data <- tidyr::unnest(res, "tomtom")
identical(nest_tomtom(data), res)
}
}
