% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{runMeme}
\alias{runMeme}
\title{Identify motifs with MEME}
\usage{
runMeme(
  input,
  control = NA,
  outdir = "auto",
  alph = "dna",
  combined_sites = FALSE,
  silent = TRUE,
  meme_path = NULL,
  ...
)
}
\arguments{
\item{control}{(default: NA), must pass `objfun` to `...` if using a control set}

\item{alph}{one of c("dna", "rna", "protein") or path to alphabet file (default: "dna")}

\item{...}{}
}
\value{

}
\description{
NOTE: add note about the hsfrac issue
}
\details{
# Citation
}
