% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomtom.R
\name{runTomTom}
\alias{runTomTom}
\title{Run TomTom on target motifs}
\usage{
runTomTom(
  input,
  database = NULL,
  outdir = "auto",
  thresh = 10,
  min_overlap = 5,
  dist = "pearson",
  evalue = T,
  meme_path = NULL,
  ...
)
}
\arguments{
\item{input}{path to .meme format file of motifs or list of universalmotifs}

\item{database}{path to .meme format file to use as reference database (or list of universalmotifs)}

\item{outdir}{directory to store tomtom results (will be overwritten if
exists). Default: location of input fasta file, or temporary location if using universalmotif list input.}

\item{thresh}{report matches less than or equal to this value. If evalue =
TRUE (default), set an e-value threshold (default = 10). If evalue = FALSE,
set a value between 0-1 (default = 0.5).}

\item{min_overlap}{only report matches that overlap by this value or more,
unless input motif is shorter, in which case the shorter length is used as
the minimum value}

\item{dist}{distance metric. Valid arguments: `allr | ed | kullback | pearson | sandelin | blic1 | blic5 | llr1 | llr5`.
Default: `pearson`.}

\item{evalue}{whether to use E-value as significance threshold (default:
`TRUE`). If evalue = FALSE, uses *q-value* instead.}

\item{meme_path}{path to "meme/bin/" (optional). If unset, will check R
environment variable "MEME_DB (set in `.Renviron`), or option
"meme_db" (set with `option(meme_db = "path/to/meme/bin")`)}

\item{...}{additional flags passed to tomtom using {dotargs} formating (see
[tomtom commandline reference](http://meme-suite.org/doc/tomtom.html?man_type=web) for details)}
}
\value{
data.frame of match results. Contains `match_motif` column of
  `universalmotif` objects with the matched PWM from the database. If no
  matches are returned, `tomtom` and `best_match_motif` columns will be set
  to `NA` and a message indicating this will print.
}
\description{
Run TomTom on target motifs
}
\examples{
\dontrun{
runTomTom("searchmotifs.meme", "jasparMotifs.meme")
}
}
