% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_utils.R
\name{update_motifs}
\alias{update_motifs}
\title{Update the \code{motif} column to data.frame values}
\usage{
update_motifs(.data)
}
\arguments{
\item{.data}{data.frame with \code{motif} column}
}
\value{
.data where \code{motif} column has been updated to reflect the values
from columns sharing names with unprotected universalmotif slots. Names of
\code{motif} list are updated to reflect name.
}
\description{
Update the \code{motif} column to data.frame values
}
\details{
\strong{NOTE} that \code{consensus}, \code{alphabet}, \code{multifreq}, and \code{icscore} are protected columns and
cannot be updated. Any changes to these columns will be \strong{ignored} when updating motifs.
\subsection{Table of values updated}{

TODO: change this table to be all universalmotif.dataframe names, w/ second column of "updated: yes/no".
| \code{motif}  | \code{data.frame} |
|:--------:|:------------:|
| name     | name         |
| altname  | altname      |
| family   | family       |
| organism | organism     |
| strand   | strand       |
| nsites   | nsites       |
| bkgsites | bkgsites     |
| pval     | pval         |
| qval     | qval         |
| eval     | eval         |
}
}
\examples{
motif <- universalmotif::create_motif()
df <- as_universalmotif_dataframe(motif)
df <- dplyr::mutate(df, id = "newName")
df <- update_motifs(df)
# renamed motif
df$motif
}
