% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{view_tomtom_hits}
\alias{view_tomtom_hits}
\title{Compare top tomtom hits to original motif}
\usage{
view_tomtom_hits(results, top_n = "all")
}
\arguments{
\item{results}{results data.frame from runTomTom}

\item{top_n}{number of matched motifs to return in plot (default: "all")}
}
\value{
plot of input motif vs the top n number of tomtom matched motifs
}
\description{
Although TomTom does a good job of matching unknown motifs to known motifs,
sometimes the top hit is not the correct assignment. It can be useful to
manually inspect the hits. This function provides a quick utility to compare
matches.
}
\examples{
\dontrun{
results <- runTomTom(motifs)
# show top 3 hits
view_tomtom_hits(results, top_n = 3)

}
}
